Move PartitionsStatus to splitter-core
