+++ java/clients/pom.xml
@@ -170,6 +170,13 @@
         </dependency>
         <dependency>
             <groupId>sleeper</groupId>
+            <artifactId>splitter-core</artifactId>
+            <version>${project.parent.version}</version>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>sleeper</groupId>
             <artifactId>common-task</artifactId>
             <version>${project.parent.version}</version>
             <type>test-jar</type>
