+++ java/clients/src/main/java/sleeper/clients/status/report/partitions/PartitionsStatusReporter.java
@@ -22,6 +22,8 @@ import sleeper.clients.util.table.TableRow;
 import sleeper.clients.util.table.TableWriterFactory;
 import sleeper.core.partition.Partition;
 import sleeper.core.schema.Field;
+import sleeper.splitter.status.PartitionStatus;
+import sleeper.splitter.status.PartitionsStatus;
 
 import java.io.PrintStream;
