+++ java/splitter/splitter-core/src/main/java/sleeper/splitter/status/PartitionStatus.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.clients.status.report.partitions;
+package sleeper.splitter.status;
 
 import sleeper.configuration.properties.table.TableProperties;
 import sleeper.core.partition.Partition;
