+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/partitioning/AwsPartitionReportDriver.java
@@ -16,9 +16,9 @@
 
 package sleeper.systemtest.drivers.partitioning;
 
-import sleeper.clients.status.report.partitions.PartitionsStatus;
 import sleeper.clients.status.report.partitions.PartitionsStatusReporter;
 import sleeper.core.statestore.StateStoreException;
+import sleeper.splitter.status.PartitionsStatus;
 import sleeper.systemtest.dsl.instance.SystemTestInstanceContext;
 import sleeper.systemtest.dsl.reporting.PartitionReportDriver;
 import sleeper.systemtest.dsl.reporting.SystemTestReport;
